refactor: modularize account balances table
