+++ frontend/app/src/modules/accounts/table/index.ts
@@ -1,0 +1,1 @@
+export { default as AccountBalancesTable } from './AccountBalancesTable.vue';
