+++ frontend/app/src/modules/accounts/table/types.ts
@@ -1,0 +1,3 @@
+import type { BlockchainAccountBalance } from '@/types/blockchain/accounts';
+
+export type AccountDataRow<T extends BlockchainAccountBalance = BlockchainAccountBalance> = T & { id: string };
